Move some stuff for reuse
