+++ tgui/packages/tgui/interfaces/programs/NtosWord.js
@@ -1,188 +1,7 @@
 import { useBackend, useSharedState } from "tgui/backend";
 import { Box, Button, Flex, Icon, Input, LabeledList, Modal, Section, Stack, Tabs, TextArea } from "tgui/components";
 import { NtosWindow } from "tgui/layouts";
-
-const ErrorMessageModal = (props, context) => {
-  const {
-    message,
-    onClick = () => {},
-    ...rest
-  } = props;
-
-  return (
-    <Modal
-      {...rest}
-    >
-      <Section
-        fitted
-        title="Error"
-      >
-        <Stack vertical>
-          <Stack.Item>
-            <Flex>
-              <Flex.Item>
-                <Icon
-                  name="times"
-                />
-              </Flex.Item>
-              <Flex.Item>
-                <Box>
-                  {message}
-                </Box>
-              </Flex.Item>
-            </Flex>
-          </Stack.Item>
-          <Stack.Item align="right">
-            <Button
-              content="OK"
-              onClick={() => onClick()}
-            />
-          </Stack.Item>
-        </Stack>
-      </Section>
-    </Modal>
-  );
-};
-
-const FileBrowserView = (props, context) => {
-  const {
-    currentFileName,
-    files={
-      local: [],
-      usb: null,
-    },
-    onAccept=(file) => {},
-    onCancel=() => {},
-    ...rest
-  } = props;
-
-  const DEVICES = {
-    LOCAL: "local",
-    USB: "usb",
-  };
-
-  const [
-    selectedDevice,
-    setSelectedDevice,
-  ] = useSharedState(context, "selectedDevice", DEVICES.LOCAL);
-
-  const [
-    selectedFile,
-    setSelectedFile,
-  ] = useSharedState(context, "selectedFile", null);
-
-  const changeTab = (selected) => {
-    setSelectedFile(null);
-    setSelectedDevice(selected);
-  };
-
-  const filesToDisplay = () => {
-    switch (selectedDevice) {
-      default:
-      case DEVICES.LOCAL:
-        return files.local;
-      case DEVICES.USB:
-        return files.usb;
-    }
-  };
-
-  return (
-    <Modal
-      p={0}
-      {...rest}
-    >
-      <Section
-        fill
-        title="Open document file"
-        buttons={
-          <Button
-            icon="times"
-            onClick={() => onCancel()}
-          />
-        }
-      >
-        <Stack
-          vertical
-          fill
-          justify="space-between"
-        >
-          <Stack.Item>
-            <Tabs fluid>
-              <Tabs.Tab
-                selected={selectedDevice === DEVICES.LOCAL}
-                onClick={() => changeTab(DEVICES.LOCAL)}
-              >
-                Local
-              </Tabs.Tab>
-              {!!files.usb && (
-                <Tabs.Tab
-                  selected={selectedDevice === DEVICES.USB}
-                  onClick={() => changeTab(DEVICES.USB)}
-                >
-                  USB device
-                </Tabs.Tab>
-              )}
-            </Tabs>
-          </Stack.Item>
-          <Stack.Item grow>
-            <Section fill scrollable>
-              <Stack vertical>
-                {filesToDisplay().map((entry, index) => (
-                  <Stack.Item key={index} mt={0}>
-                    <Button.Checkbox
-                      fluid
-                      lineHeight={2}
-                      disabled={entry.name === currentFileName}
-                      checked={selectedFile === entry.name}
-                      onClick={() => setSelectedFile(entry.name)}
-                    >
-                      <Flex
-                        inline
-                        width="95%"
-                        direction="row"
-                        justify="space-between"
-                      >
-                        <Flex.Item>
-                          {entry.name}
-                        </Flex.Item>
-                        <Flex.Item>
-                          {entry.size}
-                        </Flex.Item>
-                      </Flex>
-                    </Button.Checkbox>
-                  </Stack.Item>
-                ))}
-              </Stack>
-            </Section>
-          </Stack.Item>
-          <Stack.Item>
-            <Flex
-              direction="row"
-              justify="space-between"
-            >
-              <Flex.Item grow>
-                <LabeledList>
-                  <LabeledList.Item
-                    label="Selected"
-                  >
-                    {!!selectedFile && selectedFile}
-                  </LabeledList.Item>
-                </LabeledList>
-              </Flex.Item>
-              <Flex.Item>
-                <Button
-                  content="Open"
-                  disabled={!selectedFile}
-                  onClick={() => onAccept(selectedFile)}
-                />
-              </Flex.Item>
-            </Flex>
-          </Stack.Item>
-        </Stack>
-      </Section>
-    </Modal>
-  );
-};
+import { ErrorMessageModal, FileBrowserModal, FileNamePromptModal } from "tgui/interfaces/programs/components";
 
 const UnsavedWarningModal = (props, context) => {
   const {
@@ -227,61 +46,6 @@ const UnsavedWarningModal = (props, context) => {
               <Flex.Item>
                 <Button
                   width={8}
-                  content="Cancel"
-                  onClick={() => onCancel()}
-                />
-              </Flex.Item>
-            </Flex>
-          </Stack.Item>
-        </Stack>
-      </Section>
-    </Modal>
-  );
-};
-
-const FileNamePromptModal = (props, context) => {
-  const {
-    title = "Enter a file name",
-    value,
-    onCancel,
-    onAccept,
-    ...rest
-  } = props;
-
-  const [
-    currentInput,
-    setCurrentInput,
-  ] = useSharedState(context, "currentInput", value);
-
-  return (
-    <Modal
-      p={0}
-      {...rest}
-    >
-      <Section
-        title={title}
-      >
-        <Stack vertical>
-          <Stack.Item>
-            <Input
-              fluid
-              value={currentInput}
-              onInput={(e, v) => setCurrentInput(v)}
-            />
-          </Stack.Item>
-          <Stack.Item>
-            <Flex
-              direction="row"
-              justify="flex-end"
-            >
-              <Flex.Item mr={2}>
-                <Button
-                  content="Proceed"
-                  onClick={() => onAccept(currentInput)}
-                />
-              </Flex.Item>
-              <Flex.Item>
-                <Button
                   content="Cancel"
                   onClick={() => onCancel()}
                 />
@@ -431,7 +195,7 @@ export const NtosWord = (props, context) => {
           />
         )}
         {!!isFileBrowserOpen && (
-          <FileBrowserView
+          <FileBrowserModal
             width={40}
             height={35}
             files={{
