+++ tgui/packages/tgui/interfaces/programs/components/index.js
@@ -1,0 +1,3 @@
+export { ErrorMessageModal } from "./ErrorMessageModal";
+export { FileBrowserModal } from "./FileBrowserModal";
+export { FileNamePromptModal } from "./FileNamePromptModal";
